refactor: rename template_steps module to steps
Rename the template_steps directory to steps for better clarity and consistency.
