+++ src/program/setup/mod.rs
@@ -1,7 +1,7 @@
 pub mod exchange_config;
 pub mod setup_engine;
 pub mod setup_step;
-pub mod template_steps;
+pub mod steps;
 
 pub use exchange_config::ExchangeConfig;
 pub use setup_engine::{SetupEngine, SetupEngineState, StepState};
